Improved build system a little bit.
node-waf configure will now scan for libgit2 on the system. If it's
not found, it'll then check if libgit2 submodule has already been
checked out in the project directory. If it hasn't, it'll get checked
out. Then the configure script will run configuration on the internal
libgit2.

At build-time, the script will build the internal libgit2 if
necessary.