Added CreateBlobFromDisk to Repository.
Based on http://libgit2.github.com/libgit2/#v0.18.0/group/blob/git_blob_create_fromdisk

CreateBlobFromDisk reads a file specified by its path and creates a
blob. The Result of this function is the oid of the blob and not a blob
object. This is to circumvent reading the bytes of the blob into memory.